Fix `--delaytime=seconds` CLI option definition (refs #426)
The current definition of the `--delaytime` does not specify that it
requires a value (for specifying the delay in seconds). So, it is
treated as a boolean option, and has a value of `1` irrespective of the
value following the option.

This changes the definition to `--delaytime=s` so the value following
the option is correctly set.